WIP on fixing CMake-generated Xcode projects.
Sanitizer libraries fail to build from Xcode because when building for the iOS simulator they try to specify an OS X deployment target. (They also specify an OS X SDK, but that gets trounced by a later -isysroot flag that sets it to the Simulator SDK.)

Xcode already has a mechanism to deal with this: the $(SDKROOT) build setting. Try setting that instead of mucking with CFLAGS.

Unfortunately it doesn't suffice; the xcodeproj still specifies the OS X SDK. This might be a result of CMake's seemingly-incorrect default behavior of setting CMAKE_OSX_DEPLOYMENT_TARGET to the host machine's SDK when building for Xcode, rather than just leaving it unset.