Use backtick to separate namespace from symbol name.
To eliminate parsing ambiguity, I want to use a symbol other than period to separate the namespace portion from the primary identifier naming an Objective-C symbol.

My first instict was to go with hash ('#'), but because dot syntax can be used with classes, this could conflict with preprocessor directives if a line began with a dot-syntax expression involving a class in an anonymous namespace (which is something I'm planning to add).